# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel and Swagger",
        "description": "Getting started with Laravel and Swagger",
        "termsOfService": "",
        "contact": {
            "email": "an.nguyenbku1999@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/task/add": {
            "post": {
                "tags": [
                    "Task"
                ],
                "description": "add task",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "SessionId",
                        "in": "query",
                        "description": "Enter SessionId:",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Enter Title Task:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Enter description:",
                        "type": "string"
                    },
                    {
                        "name": "urlAvatar",
                        "in": "query",
                        "description": "Enter urlAvatar:",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/task/edit": {
            "post": {
                "tags": [
                    "Task"
                ],
                "description": "edit task",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "SessionId",
                        "in": "query",
                        "description": "Enter SessionId:",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "TaskId",
                        "in": "query",
                        "description": "Enter TaskId:",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Enter Title Task:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Enter description:",
                        "type": "string"
                    },
                    {
                        "name": "urlAvatar",
                        "in": "query",
                        "description": "Enter urlAvatar:",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/task/delete": {
            "post": {
                "tags": [
                    "Task"
                ],
                "description": "delete task",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TaskId",
                        "in": "query",
                        "description": "Enter Task ID:",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/task/addMembersListToTask": {
            "post": {
                "tags": [
                    "Task"
                ],
                "description": "add Members List To Task",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TaskId",
                        "in": "query",
                        "description": "Enter Task ID:",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "UserIdList",
                        "in": "query",
                        "description": "Enter User Member List like [1,2,3,...]",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/task/removeMember": {
            "post": {
                "tags": [
                    "Task"
                ],
                "description": "remove Member from Task",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TaskId",
                        "in": "query",
                        "description": "Enter Task ID:",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "UserId",
                        "in": "query",
                        "description": "Enter User ID Member do you want to remove:",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
...
    },
...
}